+++ src/common/object.h
@@ -19,7 +19,7 @@ typedef struct {
 	short  nsargs;			/* # of string arguments */
 	short  nfargs;			/* # of real arguments */
 	char  **sarg;			/* string arguments */
-	double  *farg;			/* real arguments */
+	FLOAT  *farg;			/* real arguments */
 #ifdef  IARGS
 	short  niargs;			/* # of integer arguments */
 	long  *iarg;			/* integer arguments */
